Remove old HomeViewModelImpl
