--- feature/home/presentation/src/main/java/com/tfandkusu/template/viewmodel/home/HomeViewModelV2Impl.kt
@@ -1,19 +1,0 @@
-package com.tfandkusu.template.viewmodel.home
-
-import com.tfandkusu.template.viewmodel.ReduxViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class HomeViewModelV2Impl @Inject constructor(
-    actionCreator: HomeActionCreator,
-    reducer: HomeReducer
-) : ReduxViewModel<HomeEvent, HomeAction, HomeState, HomeEffect>(
-    actionCreator,
-    reducer
-),
-    HomeViewModel {
-    override fun createDefaultState(): HomeState {
-        return HomeState()
-    }
-}
